Added an item model to the Tablet item and attempted to implement a custom energy bar renderer.
I finally fixed the issue in TechAscendItemModels (I was registering the Tablet as having a block parent to pull the texture from). I added an EnergyBarRenderer class under client/rendering/gui and it somewhat succesfully renders a bar depending on how much energy an item has. I say somewhat because it only works for the hotbar; for some reason shifts a pixel up when you are looking at a block; and will get hidden behind the item once you open the inventory for the first time after starting the game. There probably is an easier and better way to implement a custom durability-like bar, should look at pneumaticraft or something like it that implements a similar bar.